Add files via upload
Python code using OpenCV to track the path for equipment and machinery in an industrial environment. Keeps track of the equipment, including its path and displays the amount of time it took to return the equipment to original location. Also displays the total number of times the equipment was moved from its base in a day